Add Criterion Benchmarks (#48)
* MIN_LOG_MEMORY_SIZE

* Update Cargo.lock and Cargo.toml to add new dependencies and benchmarks; refactor rec_aggregation module to expose recursion and xmss_aggregate. Introduce examples module in the library and adjust main.rs to use updated paths.

* Add benchmark for Poseidon2 implementation

This commit introduces a new benchmark for the Poseidon2 algorithm, utilizing the Criterion library for performance measurement. The benchmark is configured with specific parameters for logging and security assumptions, and it measures the prover time of the Poseidon2 configuration. The new file is located in `benches/poseidon2.rs`.

* Add recursion benchmark using Criterion

This commit introduces a new benchmark for the recursion functionality in the `rec_aggregation` module, utilizing the Criterion library for performance measurement. The benchmark is configured with specific parameters for sample size, measurement time, and warm-up time, and it measures the proving time of the recursion implementation. The new benchmark file is located in `benches/recursion.rs`, and the recursion logic has been refactored to support this benchmarking. Additionally, constants have been introduced for better readability and maintainability.

* Add benchmark for XMSS implementation

This commit introduces a new benchmark for the XMSS algorithm in the `benches/xmss.rs` file, utilizing the Criterion library for performance measurement. The benchmark is configured with specific parameters for sample size, measurement time, and warm-up time, and it measures the proving time of the XMSS implementation. Additionally, the `xmss_aggregate.rs` file has been refactored to include a new `XmssBenchStats` struct for better organization of benchmarking results and to streamline the benchmarking process. The test function for XMSS aggregation has also been updated to utilize the new benchmarking structure.

---------

Co-authored-by: Tom Wambsgans <TomWambsgans@users.noreply.github.com>
Co-authored-by: Tom Wambsgans <101908066+TomWambsgans@users.noreply.github.com>